feat : 고정지출 CRUD 구현 완료 - 목록 조회 및 단건 조회 로직 구현 - 생성, 수정, 삭제 로직 구현

+++ server/back/src/main/java/com/barcode/honeykeep/fixedexpense/repository/FixedExpenseRepository.java
@@ -5,7 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FixedExpenseRepository extends JpaRepository<FixedExpense, Integer> {
+public interface FixedExpenseRepository extends JpaRepository<FixedExpense, Long> {
 
     List<FixedExpense> findByUserId(Integer userId);
+
 }
